01-introduction-to-data-engineering/sources/apis/pyproject.toml
[tool.poetry]
name = "apis"
version = "0.1.0"
description = ""
authors = ["zkan <zkan.cs@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.2"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.6"
django-environ = "^0.10.0"
django-filter = "^23.1"
requests = "^2.28.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
